+++ src/Jira/Issues/Walker.php
@@ -27,7 +27,7 @@ namespace chobie\Jira\Issues;
 
 use chobie\Jira\Api;
 
-class Walker implements \Iterator, \Countable
+class Walker implements \IteratorAggregate
 {
 
 	/**
@@ -45,60 +45,11 @@ class Walker implements \Iterator, \Countable
 	protected $jql = null;
 
 	/**
-	 * Offset.
-	 *
-	 * @var integer
-	 */
-	protected $offset = 0;
-
-	/**
-	 * Current record index.
-	 *
-	 * @var integer
-	 */
-	protected $current = 0;
-
-	/**
-	 * Total issue count.
-	 *
-	 * @var integer
-	 */
-	protected $total = null;
-
-	/**
-	 * Issue count on current page.
-	 *
-	 * @var integer
-	 */
-	protected $max = 0;
-
-	/**
-	 * Index of issue in issue list (across all issue pages).
-	 *
-	 * @var integer
-	 */
-	protected $startAt = 0;
-
-	/**
 	 * Issues per page.
 	 *
 	 * @var integer
 	 */
 	protected $perPage = 50;
-
-	/**
-	 * Was JQL executed.
-	 *
-	 * @var boolean
-	 */
-	protected $executed = false;
-
-	/**
-	 * Result.
-	 *
-	 * @var array
-	 */
-	protected $issues = array();
 
 	/**
 	 * List of fields to query.
@@ -144,143 +95,53 @@ class Walker implements \Iterator, \Countable
 	}
 
 	/**
-	 * Return the current element.
+	 * @inheritDoc
 	 *
-	 * @return mixed Can return any type.
-	 * @link   http://php.net/manual/en/iterator.current.php
-	 */
-	public function current()
-	{
-		if ( is_callable($this->callback) ) {
-			$tmp = $this->issues[$this->offset];
-			$callback = $this->callback;
-
-			return $callback($tmp);
-		}
-
-		return $this->issues[$this->offset];
-	}
-
-	/**
-	 * Move forward to next element.
-	 *
-	 * @return void Any returned value is ignored.
-	 * @link   http://php.net/manual/en/iterator.next.php
-	 */
-	public function next()
-	{
-		$this->offset++;
-	}
-
-	/**
-	 * Return the key of the current element.
-	 *
-	 * @return mixed scalar on success, or null on failure.
-	 * @link   http://php.net/manual/en/iterator.key.php
-	 */
-	public function key()
-	{
-		if ( $this->startAt > 0 ) {
-			return $this->offset + (($this->startAt - 1) * $this->perPage);
-		}
-
-		return 0;
-	}
-
-	/**
-	 * Checks if current position is valid.
-	 *
-	 * @return boolean The return value will be casted to boolean and then evaluated.
-	 *                 Returns true on success or false on failure.
 	 * @throws \Exception When "Walker::push" method wasn't called.
 	 * @throws Api\UnauthorizedException When it happens.
-	 * @link   http://php.net/manual/en/iterator.valid.php
 	 */
-	public function valid()
+	public function getIterator()
 	{
 		if ( $this->jql === null ) {
 			throw new \Exception('you have to call Jira_Walker::push($jql, $fields) at first');
 		}
 
-		if ( !$this->executed ) {
-			try {
-				$result = $this->api->search($this->getQuery(), $this->key(), $this->perPage, $this->fields);
+		$next_page_token = null;
+		$jql = $this->getQuery();
 
-				$this->setResult($result);
-				$this->executed = true;
+		try {
+			while ( true ) {
+				$result = $this->api->search($jql, $next_page_token, $this->perPage, $this->fields);
 
-				if ( $result->getTotal() == 0 ) {
-					return false;
+				if ( $result->getIssuesCount() === 0 ) {
+					return;
 				}
 
-				return true;
-			}
-			catch ( Api\UnauthorizedException $e ) {
-				throw $e;
-			}
-			catch ( \Exception $e ) {
-				error_log($e->getMessage());
+				foreach ( $result->getIssues() as $issue ) {
+					if ( is_callable($this->callback) ) {
+						$callback = $this->callback;
 
-				return false;
+						yield $callback($issue);
+					}
+					else {
+						yield $issue;
+					}
+				}
+
+				$raw_result = $result->getResult();
+				$next_page_token = array_key_exists('nextPageToken', $raw_result) ? $raw_result['nextPageToken'] : null;
+
+				if ( empty($next_page_token) ) {
+					break;
+				}
 			}
 		}
-		else {
-			if ( $this->offset >= $this->max && $this->key() < $this->total ) {
-				try {
-					$result = $this->api->search($this->getQuery(), $this->key(), $this->perPage, $this->fields);
-					$this->setResult($result);
-
-					return true;
-				}
-				catch ( Api\UnauthorizedException $e ) {
-					throw $e;
-				}
-				catch ( \Exception $e ) {
-					error_log($e->getMessage());
-
-					return false;
-				}
-			}
-			else {
-				if ( ($this->startAt - 1) * $this->perPage + $this->offset < $this->total ) {
-					return true;
-				}
-
-				return false;
-			}
+		catch ( Api\UnauthorizedException $e ) {
+			throw $e;
 		}
-	}
-
-	/**
-	 * Rewind the Iterator to the first element.
-	 *
-	 * @return void Any returned value is ignored.
-	 * @link   http://php.net/manual/en/iterator.rewind.php
-	 */
-	public function rewind()
-	{
-		$this->offset = 0;
-		$this->startAt = 0;
-		$this->current = 0;
-		$this->max = 0;
-		$this->total = null;
-		$this->executed = false;
-		$this->issues = array();
-	}
-
-	/**
-	 * Count elements of an object.
-	 *
-	 * @return integer The custom count as an integer.
-	 * @link   http://php.net/manual/en/countable.count.php
-	 */
-	public function count()
-	{
-		if ( $this->total === null ) {
-			$this->valid();
+		catch ( \Exception $e ) {
+			error_log($e->getMessage());
 		}
-
-		return $this->total;
 	}
 
 	/**
@@ -299,22 +160,6 @@ class Walker implements \Iterator, \Countable
 		else {
 			throw new \Exception('passed argument is not callable');
 		}
-	}
-
-	/**
-	 * Sets result.
-	 *
-	 * @param Api\Result $result Result.
-	 *
-	 * @return void
-	 */
-	protected function setResult(Api\Result $result)
-	{
-		$this->total = $result->getTotal();
-		$this->offset = 0;
-		$this->max = $result->getIssuesCount();
-		$this->issues = $result->getIssues();
-		$this->startAt++;
 	}
 
 	/**