+++ CHANGELOG.md
@@ -10,9 +10,11 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 - The `Api::downloadAttachment` method now throws an exception, when attempting to download from a non-Jira website by [@aik099] (#240).
 - The `$params` argument of the `Api::getWorklogs` method is now optional by [@aik099] (#244).
 - The `$params` argument of the `Api::getTransitions` method is now optional by [@aik099] (#244).
+- The `$start_at = 0` argument of `Api::search` method was renamed into the `$next_page_token = null` to reflect underlying Jira API changes by [@aik099] (#248).
+- Improved performance of `Result::getIssuesCount` method by [@aik099] (#248).
 
 ### Removed
-...
+- The `Api::search` method and the `Walker` class no longer return total issue count, because underlaying Jira API doesn't support this [@aik099] (#248).
 
 ### Fixed
 - The `CurlClient` in combination with cURL version < 7.33.0 was getting `426 Upgrade Required` error on any request to the Atlassian Cloud servers by [@aik099] (#239).